requirements.txt
bibtexparser
pandas
requests
scholarly
openpyxl
groq
python-dotenv
